requirements.txt

asgiref==3.3.4
click==7.1.2
Django==3.1.7
gunicorn==20.1.0
joblib==1.0.1
nltk==3.6.2
numpy==1.20.2
pytz==2021.1
regex==2021.4.4
scikit-learn==0.24.1
scipy==1.6.3
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.60.0
whitenoise==5.2.0
google-cloud-vision==3.4.5
google-auth==2.29.0
google-generativeai==0.4.1
Pillow==8.2.0
python-dotenv==1.0.1
